begin work on stats

+++ Assets/Scripts/Entity/IEntityActions.cs
@@ -5,7 +5,7 @@ namespace DefaultNamespace
     public interface IEntityActions
     {
         void Move(Vector3 destination);
-        void Damage(float amt);
+        void Damage(int amt);
         void Die();
         void Attack(Entity entity);
     }
